+++ NEWS.md
@@ -102,6 +102,8 @@
       * GCMM images are not supported yet because there doesn't seem to
         be any documentation on how to create one.
     * Added China and Taiwan region codes.
+    * Added support for standalone Wii partitions. These are commonly used
+      with the NKit tool to restore full Wii disc images.
   * NES: Added support for the internal footer present in some ROMs. This
     footer is used for, among other things, FamicomBox.
   * Nintendo3DS: Split the SMDH code into a separate parser. This should